fix(cosi): make grant idempotent, reject key reuse
Align DriverGrantBucketAccess with Ceph-style isolation: deterministic
secrets, never rotate existing users, and return AlreadyExists when
preferredAccessKey is claimed by another BucketAccess.
