+++ crates/cosi-driver/Cargo.toml
@@ -1,9 +1,8 @@
 [package]
-name = "cosi-driver"
+name = "rustfs-cosi-driver"
 version = "0.1.0"
 edition = "2024"
 license = "Apache-2.0"
-description = "RustFS Container Object Storage Interface (COSI) v1alpha1 driver"
 publish = false
 
 [[bin]]
@@ -11,29 +10,22 @@ name = "rustfs-cosi-driver"
 path = "src/main.rs"
 
 [dependencies]
-hex = "0.4"
-k8s-openapi = { version = "0.26.1", features = ["v1_30"] }
-kube = { version = "2.0.1", features = ["client", "rustls-tls"] }
-prost = "0.13"
-rustfs-admin = { path = "../rustfs-admin" }
-sha2 = "0.10"
-snafu = { version = "0.8.9", features = ["futures"] }
+operator = { path = "../.." }
 tokio = { version = "1.49.0", features = ["rt", "rt-multi-thread", "macros", "fs", "signal", "net"] }
 tokio-stream = { version = "0.1", features = ["net"] }
 tonic = "0.12"
-tracing = "0.1.44"
-tracing-subscriber = { version = "0.3.22", features = ["env-filter"] }
+prost = "0.13"
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+kube = { version = "2.0.1", features = ["client", "rustls-tls"] }
+k8s-openapi = { version = "0.26.1", features = ["v1_30"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+sha2 = "0.10"
+hex = "0.4"
+thiserror = "2"
+rustls = { version = "0.23", default-features = false, features = ["ring"] }
 
 [build-dependencies]
-protoc-bin-vendored = "3"
 tonic-build = "0.12"
-
-[dev-dependencies]
-tokio = { version = "1.49.0", features = ["rt", "macros"] }
-
-[lints.rust]
-unused_variables = "allow"
-
-[lints.clippy]
-unwrap_used = "deny"
-expect_used = "deny"
+protoc-bin-vendored = "3"
