+++ crates/cosi-driver/build.rs
@@ -1,24 +1,13 @@
-// Copyright 2025 RustFS Team
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let protoc = protoc_bin_vendored::protoc_bin_path()?;
-    // SAFETY: build scripts run single-threaded before compilation.
-    unsafe {
-        std::env::set_var("PROTOC", protoc);
+    // Prefer PATH `protoc` (Dockerfile installs protobuf-compiler); fall back to vendored.
+    if std::env::var_os("PROTOC").is_none()
+        && let Ok(protoc) = protoc_bin_vendored::protoc_bin_path()
+    {
+        // SAFETY: build script is single-threaded before codegen.
+        unsafe {
+            std::env::set_var("PROTOC", protoc);
+        }
     }
-
     tonic_build::configure()
         .build_server(true)
         .build_client(false)
