+++ crates/cosi-driver/proto/cosi.proto
@@ -1,78 +1,43 @@
-// Code generated by make; DO NOT EDIT.
 syntax = "proto3";
 package cosi.v1alpha1;
 
+option go_package = "sigs.k8s.io/container-object-storage-interface/proto;cosi";
 
 service Identity {
-    // This call is meant to retrieve the unique provisioner Identity.
-    // This identity will have to be set in BucketClaim.DriverName field in order to invoke this specific provisioner.
     rpc DriverGetInfo (DriverGetInfoRequest) returns (DriverGetInfoResponse) {}
 }
 
 service Provisioner {
-    // This call is made to create the bucket in the backend.
-    // This call is idempotent
-    //    1. If a bucket that matches both name and parameters already exists, then OK (success) must be returned.
-    //    2. If a bucket by same name, but different parameters is provided, then the appropriate error code ALREADY_EXISTS must be returned.
     rpc DriverCreateBucket (DriverCreateBucketRequest) returns (DriverCreateBucketResponse) {}
-    // This call is made to delete the bucket in the backend.
-    // If the bucket has already been deleted, then no error should be returned.
     rpc DriverDeleteBucket (DriverDeleteBucketRequest) returns (DriverDeleteBucketResponse) {}
-
-    // This call grants access to an account. The account_name in the request shall be used as a unique identifier to create credentials.
-    // The account_id returned in the response will be used as the unique identifier for deleting this access when calling DriverRevokeBucketAccess.
     rpc DriverGrantBucketAccess (DriverGrantBucketAccessRequest) returns (DriverGrantBucketAccessResponse);
-    // This call revokes all access to a particular bucket from a principal.
     rpc DriverRevokeBucketAccess (DriverRevokeBucketAccessRequest) returns (DriverRevokeBucketAccessResponse);
 }
 
-// S3SignatureVersion is the version of the signing algorithm for all s3 requests
 enum S3SignatureVersion {
     UnknownSignature = 0;
-    // S3V2, Signature version v2
     S3V2 = 1;
-    // S3V4, Signature version v4
     S3V4 = 2;
-}
-
-enum AnonymousBucketAccessMode {
-    UnknownBucketAccessMode = 0;
-    // Default, disallow uncredentialed access to the backend storage.
-    Private = 1;
-    // Read only, uncredentialed users can call ListBucket and GetObject.
-    ReadOnly = 2;
-    // Write only, uncredentialed users can only call PutObject.
-    WriteOnly = 3;
-    // Read/Write, uncredentialed users can read objects as well as PutObject.
-    ReadWrite = 4;
 }
 
 enum AuthenticationType {
     UnknownAuthenticationType = 0;
-    // Default, KEY based authentication.
     Key = 1;
-    // Storageaccount based authentication.
     IAM = 2;
 }
 
 message S3 {
-    // region denotes the geographical region where the S3 server is running
     string region = 1;
-    // signature_version denotes the signature version for signing all s3 requests
     S3SignatureVersion signature_version = 2;
 }
 
 message AzureBlob {
-    // storage_account is the id of the azure storage account
     string storage_account = 1;
 }
 
 message GCS {
-    // private_key_name denotes the name of the private key in the storage backend
     string private_key_name = 1;
-    // project_id denotes the name of the project id in the storage backend
     string project_id = 2;
-    // service_account denotes the name of the service account in the storage backend
     string service_account = 3;
 }
 
@@ -85,112 +50,48 @@ message Protocol {
 }
 
 message CredentialDetails {
-    // map of the details in the secrets for the protocol string
     map<string, string> secrets = 1;
 }
 
-message DriverGetInfoRequest {
-    // Intentionally left blank
-}
+message DriverGetInfoRequest {}
 
 message DriverGetInfoResponse {
-    // This field is REQUIRED
-    // The name MUST follow domain name notation format
-    // (https://tools.ietf.org/html/rfc1035#section-2.3.1). It SHOULD
-    // include the plugin's host company name and the plugin name,
-    // to minimize the possibility of collisions. It MUST be 63
-    // characters or less, beginning and ending with an alphanumeric
-    // character () with dashes (-), dots (.), and
-    // alphanumerics between.
     string name = 1;
 }
 
 message DriverCreateBucketRequest {
-    // This field is REQUIRED
-    // name specifies the name of the bucket that should be created.
     string name = 1;
-
-    // This field is OPTIONAL
-    // The caller should treat the values in parameters as opaque.
-    // The receiver is responsible for parsing and validating the values.
     map<string,string> parameters = 2;
 }
 
 message DriverCreateBucketResponse {
-    // bucket_id returned here is expected to be the globally unique
-    // identifier for the bucket in the object storage provider.
     string bucket_id = 1;
-
-    // bucket_info returned here stores the data specific to the
-    // bucket required by the object storage provider to connect to the bucket.
     Protocol bucket_info = 2;
 }
 
 message DriverDeleteBucketRequest {
-    // This field is REQUIRED
-    // bucket_id is a globally unique identifier for the bucket
-    // in the object storage provider
     string bucket_id = 1;
-
-    // This field is OPTIONAL
-    // The caller should treat the values in delete_context as opaque.
-    // The receiver is responsible for parsing and validating the values.
     map<string,string> delete_context = 2;
 }
 
-message DriverDeleteBucketResponse {
-    // Intentionally left blank
-}
+message DriverDeleteBucketResponse {}
 
 message DriverGrantBucketAccessRequest {
-    // This field is REQUIRED
-    // bucket_id is a globally unique identifier for the bucket
-    // in the object storage provider
     string bucket_id = 1;
-
-    // This field is REQUIRED
-    // name field is used to define the name of the bucket access object.
     string name = 2;
-
-    // This field is REQUIRED
-    // Requested authentication type for the bucket access.
-    // Supported authentication types are KEY or IAM.
     AuthenticationType authentication_type = 3;
-
-    // This field is OPTIONAL
-    // The caller should treat the values in parameters as opaque.
-    // The receiver is responsible for parsing and validating the values.
     map<string,string> parameters = 4;
 }
 
 message DriverGrantBucketAccessResponse {
-    // This field is REQUIRED
-    // This is the account_id that is being provided access. This will
-    // be required later to revoke access.
     string account_id = 1;
-
-    // This field is REQUIRED
-    // Credentials supplied for accessing the bucket ex: aws access key id and secret, etc.
     map<string, CredentialDetails> credentials = 2;
 }
 
 message DriverRevokeBucketAccessRequest {
-    // This field is REQUIRED
-    // bucket_id is a globally unique identifier for the bucket
-    // in the object storage provider.
     string bucket_id = 1;
-
-    // This field is REQUIRED
-    // This is the account_id that is having its access revoked.
     string account_id = 2;
-
-    // This field is OPTIONAL
-    // The caller should treat the values in revoke_access_context as opaque.
-    // The receiver is responsible for parsing and validating the values.
     map<string,string> revoke_access_context = 3;
 }
 
-message DriverRevokeBucketAccessResponse {
-    // Intentionally left blank
-}
-
+message DriverRevokeBucketAccessResponse {}
