+++ crates/cosi-driver/src/driver.rs
@@ -1,30 +1,17 @@
-// Copyright 2025 RustFS Team
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//! gRPC Identity and Provisioner servers for COSI v1alpha1.
-
-#![allow(clippy::result_large_err)]
+//! COSI Identity + Provisioner gRPC services.
 
 use std::collections::HashMap;
 
-use sha2::{Digest, Sha256};
+use kube::Client;
+use operator::sts::rustfs_client::{CreateBucketResult, RustfsAdminClient};
 use tonic::{Request, Response, Status};
-use tracing::info;
+use tracing::{error, info};
 
-use crate::backend::{BackendError, BackendFactory};
-use crate::parameters::{BackendParameters, ParameterError};
-use crate::policy::{bucket_policy_document, policy_name_for};
+use crate::backend::{BackendError, admin_client_from_params};
+use crate::parameters::{
+    BackendParameters, DRIVER_NAME, bucket_policy_document_for, credentials_for_account,
+    grant_owner_policy_document, grant_owner_policy_name, policy_name_for,
+};
 use crate::proto::cosi::v1alpha1::{
     AuthenticationType, CredentialDetails, DriverCreateBucketRequest, DriverCreateBucketResponse,
     DriverDeleteBucketRequest, DriverDeleteBucketResponse, DriverGetInfoRequest,
@@ -33,78 +20,148 @@ use crate::proto::cosi::v1alpha1::{
     S3SignatureVersion, identity_server::Identity, provisioner_server::Provisioner,
 };
 
-pub const DRIVER_NAME: &str = "rustfs.objectstorage.k8s.io";
-
-/// Deterministic secret so DriverGrantBucketAccess is idempotent across sidecar retries.
-fn credentials_for_account(account_id: &str) -> String {
-    let digest = Sha256::digest(format!("rustfs-cosi-v1:{account_id}").as_bytes());
-    hex::encode(digest)
+#[derive(Clone)]
+pub struct Driver {
+    kube: Client,
 }
 
-#[cfg(test)]
-mod credential_tests {
-    use super::credentials_for_account;
-
-    #[test]
-    fn credentials_are_deterministic_and_long_enough() {
-        let a = credentials_for_account("ba-test-uid");
-        let b = credentials_for_account("ba-test-uid");
-        assert_eq!(a, b);
-        assert!(a.len() >= 8);
-        assert_ne!(a, credentials_for_account("other-account"));
+impl Driver {
+    pub fn new(kube: Client) -> Self {
+        Self { kube }
     }
 }
 
-pub struct IdentityService {
-    pub name: String,
+fn map_backend(err: BackendError) -> Status {
+    error!(error = %err, "backend error");
+    Status::internal(err.to_string())
+}
+
+fn map_params(err: crate::parameters::ParameterError) -> Status {
+    Status::invalid_argument(err.to_string())
+}
+
+fn map_admin(err: operator::sts::rustfs_client::RustfsClientError) -> Status {
+    error!(error = %err, "rustfs admin error");
+    Status::internal(err.to_string())
+}
+
+fn credential_map(
+    access_key: &str,
+    secret_key: &str,
+    params: &BackendParameters,
+    policy_buckets: &[String],
+) -> HashMap<String, String> {
+    let mut secrets = HashMap::new();
+    secrets.insert("accessKeyID".to_string(), access_key.to_string());
+    secrets.insert("accessSecretKey".to_string(), secret_key.to_string());
+    secrets.insert("AWS_ACCESS_KEY_ID".to_string(), access_key.to_string());
+    secrets.insert("AWS_SECRET_ACCESS_KEY".to_string(), secret_key.to_string());
+    secrets.insert("ACCESSKEY".to_string(), access_key.to_string());
+    secrets.insert("SECRETKEY".to_string(), secret_key.to_string());
+    secrets.insert("endpoint".to_string(), params.endpoint.clone());
+    secrets.insert("region".to_string(), params.region.clone());
+    secrets.insert(
+        "BUCKETS".to_string(),
+        params
+            .buckets
+            .clone()
+            .unwrap_or_else(|| policy_buckets.join(",")),
+    );
+    secrets
+}
+
+fn user_owns_grant(
+    policy_names: &[String],
+    owner_policy: &str,
+    account_id: &str,
+    grant_name: &str,
+) -> bool {
+    // Default Ceph-style path: account id is the COSI grant name itself.
+    if account_id == grant_name {
+        return true;
+    }
+    policy_names.iter().any(|name| name == owner_policy)
+}
+
+async fn ensure_grant_policies(
+    client: &RustfsAdminClient,
+    access_key: &str,
+    bucket_policy_name: &str,
+    bucket_policy_doc: &str,
+    owner_policy_name: &str,
+) -> Result<(), Status> {
+    client
+        .add_canned_policy(bucket_policy_name, bucket_policy_doc)
+        .await
+        .map_err(map_admin)?;
+    client
+        .add_canned_policy(owner_policy_name, &grant_owner_policy_document())
+        .await
+        .map_err(map_admin)?;
+    client
+        .set_user_policy(
+            access_key,
+            &[
+                bucket_policy_name.to_string(),
+                owner_policy_name.to_string(),
+            ],
+        )
+        .await
+        .map_err(map_admin)?;
+    Ok(())
 }
 
 #[tonic::async_trait]
-impl Identity for IdentityService {
+impl Identity for Driver {
     async fn driver_get_info(
         &self,
         _request: Request<DriverGetInfoRequest>,
     ) -> Result<Response<DriverGetInfoResponse>, Status> {
         Ok(Response::new(DriverGetInfoResponse {
-            name: self.name.clone(),
+            name: DRIVER_NAME.to_string(),
         }))
     }
 }
 
-pub struct ProvisionerService {
-    pub backend: BackendFactory,
-}
-
 #[tonic::async_trait]
-impl Provisioner for ProvisionerService {
+impl Provisioner for Driver {
     async fn driver_create_bucket(
         &self,
         request: Request<DriverCreateBucketRequest>,
     ) -> Result<Response<DriverCreateBucketResponse>, Status> {
         let req = request.into_inner();
-        let bucket_name = req.name.trim();
-        if bucket_name.is_empty() {
+        if req.name.trim().is_empty() {
             return Err(Status::invalid_argument("bucket name is required"));
         }
-
-        let params = parse_params(&req.parameters)?;
-        let client = self
-            .backend
-            .admin_client(&params)
+        let params = BackendParameters::from_map(&req.parameters).map_err(map_params)?;
+        let client = admin_client_from_params(&self.kube, &params)
             .await
             .map_err(map_backend)?;
 
-        info!(bucket = %bucket_name, endpoint = %params.endpoint, "creating bucket");
-        client
-            .create_bucket(bucket_name, params.region.as_deref(), false)
-            .await
-            .map_err(map_admin)?;
+        let buckets = params.buckets_to_create(&req.name);
+        if buckets.is_empty() {
+            return Err(Status::invalid_argument(
+                "no buckets to create (buckets/bucketName empty or only *)",
+            ));
+        }
+        let bucket_id = params.primary_bucket_id(&req.name);
+
+        for bucket in &buckets {
+            info!(bucket = %bucket, cosi_name = %req.name, "creating bucket");
+            match client
+                .create_bucket(bucket, Some(params.region.as_str()), false)
+                .await
+                .map_err(map_admin)?
+            {
+                CreateBucketResult::Created | CreateBucketResult::AlreadyExists => {}
+            }
+        }
 
         Ok(Response::new(DriverCreateBucketResponse {
-            bucket_id: bucket_name.to_string(),
+            bucket_id,
             bucket_info: Some(Protocol {
                 r#type: Some(crate::proto::cosi::v1alpha1::protocol::Type::S3(S3 {
-                    region: params.region.unwrap_or_else(|| "us-east-1".to_string()),
+                    region: params.region,
                     signature_version: S3SignatureVersion::S3v4 as i32,
                 })),
             }),
@@ -116,21 +173,25 @@ impl Provisioner for ProvisionerService {
         request: Request<DriverDeleteBucketRequest>,
     ) -> Result<Response<DriverDeleteBucketResponse>, Status> {
         let req = request.into_inner();
-        let bucket_id = req.bucket_id.trim();
-        if bucket_id.is_empty() {
+        if req.bucket_id.trim().is_empty() {
             return Err(Status::invalid_argument("bucket_id is required"));
         }
-
-        let params = parse_params(&req.delete_context)?;
-        let client = self
-            .backend
-            .admin_client(&params)
+        let params = BackendParameters::from_map(&req.delete_context).map_err(map_params)?;
+        let client = admin_client_from_params(&self.kube, &params)
             .await
             .map_err(map_backend)?;
 
-        info!(bucket = %bucket_id, "deleting bucket");
-        client.delete_bucket(bucket_id).await.map_err(map_admin)?;
+        let buckets = params.buckets_to_create(&req.bucket_id);
+        let targets = if buckets.is_empty() {
+            vec![req.bucket_id.clone()]
+        } else {
+            buckets
+        };
 
+        for bucket in &targets {
+            info!(bucket = %bucket, "deleting bucket");
+            client.delete_bucket(bucket).await.map_err(map_admin)?;
+        }
         Ok(Response::new(DriverDeleteBucketResponse {}))
     }
 
@@ -139,12 +200,10 @@ impl Provisioner for ProvisionerService {
         request: Request<DriverGrantBucketAccessRequest>,
     ) -> Result<Response<DriverGrantBucketAccessResponse>, Status> {
         let req = request.into_inner();
-        let bucket_id = req.bucket_id.trim();
-        let account_name = req.name.trim();
-        if bucket_id.is_empty() {
+        if req.bucket_id.trim().is_empty() {
             return Err(Status::invalid_argument("bucket_id is required"));
         }
-        if account_name.is_empty() {
+        if req.name.trim().is_empty() {
             return Err(Status::invalid_argument("account name is required"));
         }
         if req.authentication_type != AuthenticationType::Key as i32
@@ -155,56 +214,80 @@ impl Provisioner for ProvisionerService {
             ));
         }
 
-        let params = parse_params(&req.parameters)?;
-        let client = self
-            .backend
-            .admin_client(&params)
+        let params = BackendParameters::from_map(&req.parameters).map_err(map_params)?;
+        let client = admin_client_from_params(&self.kube, &params)
             .await
             .map_err(map_backend)?;
 
-        let account_id = account_name.to_string();
-        let secret_key = credentials_for_account(&account_id);
-        let policy_name = policy_name_for(&account_id, bucket_id);
-        let policy_doc = bucket_policy_document(bucket_id, params.access_policy);
+        let grant_name = req.name.clone();
+        let access_key = params
+            .preferred_access_key
+            .clone()
+            .unwrap_or_else(|| grant_name.clone());
+        let secret_key = credentials_for_account(&access_key);
+        let policy_buckets = params.buckets_for_policy(&req.bucket_id);
+        let bucket_policy_name = params
+            .policy
+            .clone()
+            .unwrap_or_else(|| policy_name_for(&access_key));
+        let bucket_policy_doc = bucket_policy_document_for(&policy_buckets);
+        let owner_policy_name = grant_owner_policy_name(&grant_name);
 
         info!(
-            bucket = %bucket_id,
-            account = %account_id,
-            policy = %policy_name,
+            bucket = %req.bucket_id,
+            account = %access_key,
+            grant = %grant_name,
+            policy = %bucket_policy_name,
+            owner_policy = %owner_policy_name,
+            buckets = %policy_buckets.join(","),
             "granting bucket access"
         );
 
-        client
-            .add_canned_policy(&policy_name, &policy_doc)
-            .await
-            .map_err(map_admin)?;
-
-        if !client.user_exists(&account_id).await.map_err(map_admin)? {
-            client
-                .add_user(&account_id, &secret_key)
-                .await
-                .map_err(map_admin)?;
+        match client.get_user_info(&access_key).await.map_err(map_admin)? {
+            Some(info) => {
+                if !user_owns_grant(
+                    &info.policy_names,
+                    &owner_policy_name,
+                    &access_key,
+                    &grant_name,
+                ) {
+                    return Err(Status::already_exists(format!(
+                        "preferredAccessKey `{access_key}` is already bound to another BucketAccess; \
+                         omit preferredAccessKey or choose a unique value"
+                    )));
+                }
+                // Same grant retry (or Ceph-style account == grant name): never rotate secret.
+                ensure_grant_policies(
+                    &client,
+                    &access_key,
+                    &bucket_policy_name,
+                    &bucket_policy_doc,
+                    &owner_policy_name,
+                )
+                .await?;
+            }
+            None => {
+                client
+                    .add_user(&access_key, &secret_key)
+                    .await
+                    .map_err(map_admin)?;
+                ensure_grant_policies(
+                    &client,
+                    &access_key,
+                    &bucket_policy_name,
+                    &bucket_policy_doc,
+                    &owner_policy_name,
+                )
+                .await?;
+            }
         }
-        client
-            .set_user_policy(&account_id, std::slice::from_ref(&policy_name))
-            .await
-            .map_err(map_admin)?;
 
-        let mut secrets = HashMap::new();
-        secrets.insert("endpoint".to_string(), params.endpoint.clone());
-        secrets.insert(
-            "region".to_string(),
-            params.region.unwrap_or_else(|| "us-east-1".to_string()),
-        );
-        secrets.insert("accessKeyID".to_string(), account_id.clone());
-        secrets.insert("accessSecretKey".to_string(), secret_key);
-        secrets.insert("bucketName".to_string(), bucket_id.to_string());
-
+        let secrets = credential_map(&access_key, &secret_key, &params, &policy_buckets);
         let mut credentials = HashMap::new();
         credentials.insert("s3".to_string(), CredentialDetails { secrets });
 
         Ok(Response::new(DriverGrantBucketAccessResponse {
-            account_id,
+            account_id: access_key,
             credentials,
         }))
     }
@@ -214,65 +297,53 @@ impl Provisioner for ProvisionerService {
         request: Request<DriverRevokeBucketAccessRequest>,
     ) -> Result<Response<DriverRevokeBucketAccessResponse>, Status> {
         let req = request.into_inner();
-        let bucket_id = req.bucket_id.trim();
-        let account_id = req.account_id.trim();
-        if account_id.is_empty() {
+        if req.bucket_id.trim().is_empty() {
+            return Err(Status::invalid_argument("bucket_id is required"));
+        }
+        if req.account_id.trim().is_empty() {
             return Err(Status::invalid_argument("account_id is required"));
         }
 
-        let params = parse_params(&req.revoke_access_context)?;
-        let client = self
-            .backend
-            .admin_client(&params)
+        let params = BackendParameters::from_map(&req.revoke_access_context).map_err(map_params)?;
+        let client = admin_client_from_params(&self.kube, &params)
             .await
             .map_err(map_backend)?;
 
-        let policy_name = if bucket_id.is_empty() {
-            None
-        } else {
-            Some(policy_name_for(account_id, bucket_id))
-        };
-
-        info!(account = %account_id, bucket = %bucket_id, "revoking bucket access");
-        client.remove_user(account_id).await.map_err(map_admin)?;
-        if let Some(policy_name) = policy_name {
-            client
-                .remove_canned_policy(&policy_name)
-                .await
-                .map_err(map_admin)?;
-        }
-
+        info!(
+            bucket = %req.bucket_id,
+            account = %req.account_id,
+            "revoking bucket access"
+        );
+        client
+            .remove_user(&req.account_id)
+            .await
+            .map_err(map_admin)?;
         Ok(Response::new(DriverRevokeBucketAccessResponse {}))
     }
 }
 
-fn parse_params(params: &HashMap<String, String>) -> Result<BackendParameters, Status> {
-    BackendParameters::from_map(params).map_err(map_params)
-}
+#[cfg(test)]
+mod grant_tests {
+    use super::user_owns_grant;
 
-fn map_params(err: ParameterError) -> Status {
-    Status::invalid_argument(err.to_string())
-}
+    #[test]
+    fn same_grant_name_as_account_is_idempotent() {
+        assert!(user_owns_grant(&[], "cosi-grant-ba-1", "ba-1", "ba-1"));
+    }
 
-fn map_backend(err: BackendError) -> Status {
-    Status::failed_precondition(err.to_string())
-}
-
-fn map_admin(err: rustfs_admin::RustfsClientError) -> Status {
-    match &err {
-        rustfs_admin::RustfsClientError::UnexpectedStatus { status, .. }
-            if status.as_u16() == 409 =>
-        {
-            Status::already_exists(err.to_string())
-        }
-        rustfs_admin::RustfsClientError::InvalidPolicyName
-        | rustfs_admin::RustfsClientError::InvalidPolicyDocument
-        | rustfs_admin::RustfsClientError::InvalidCredentialValue { .. }
-        | rustfs_admin::RustfsClientError::EmptyCredentialValue { .. }
-        | rustfs_admin::RustfsClientError::MissingCredentialKey { .. }
-        | rustfs_admin::RustfsClientError::RequestBuildFailed => {
-            Status::invalid_argument(err.to_string())
-        }
-        _ => Status::internal(err.to_string()),
+    #[test]
+    fn preferred_key_requires_owner_marker() {
+        assert!(!user_owns_grant(
+            &["cosi-mlflow".to_string()],
+            "cosi-grant-ba-1",
+            "mlflow",
+            "ba-1"
+        ));
+        assert!(user_owns_grant(
+            &["cosi-mlflow".to_string(), "cosi-grant-ba-1".to_string()],
+            "cosi-grant-ba-1",
+            "mlflow",
+            "ba-1"
+        ));
     }
 }
