+++ crates/cosi-driver/src/main.rs
@@ -1,56 +1,83 @@
-// Copyright 2025 RustFS Team
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+//! RustFS COSI driver — serves Identity + Provisioner on a Unix socket.
+
+mod backend;
+mod driver;
+mod parameters;
+
+pub mod proto {
+    pub mod cosi {
+        pub mod v1alpha1 {
+            tonic::include_proto!("cosi.v1alpha1");
+        }
+    }
+}
 
 use std::path::PathBuf;
-use std::time::Duration;
 
-use cosi_driver::backend::BackendFactory;
-use cosi_driver::driver::{DRIVER_NAME, IdentityService, ProvisionerService};
-use cosi_driver::proto::cosi::v1alpha1::{
-    identity_server::IdentityServer, provisioner_server::ProvisionerServer,
-};
+use kube::Client;
 use tokio::net::UnixListener;
 use tokio_stream::wrappers::UnixListenerStream;
-use tonic::transport::Server;
-use tracing::{info, warn};
-use tracing_subscriber::EnvFilter;
+use tracing::{error, info};
 
-const DEFAULT_ENDPOINT: &str = "unix:///var/lib/cosi/cosi.sock";
-const SHUTDOWN_GRACE: Duration = Duration::from_secs(5);
+use crate::driver::Driver;
+use crate::parameters::DRIVER_NAME;
+use crate::proto::cosi::v1alpha1::{
+    identity_server::IdentityServer, provisioner_server::ProvisionerServer,
+};
+
+fn parse_unix_endpoint(raw: &str) -> Result<PathBuf, String> {
+    let trimmed = raw.trim();
+    let path = trimmed
+        .strip_prefix("unix://")
+        .ok_or_else(|| format!("unsupported COSI_ENDPOINT `{trimmed}`"))?;
+    if path.is_empty() {
+        return Err("COSI_ENDPOINT unix path is empty".to_string());
+    }
+    Ok(PathBuf::from(path))
+}
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tracing_subscriber::fmt()
-        .with_env_filter(EnvFilter::from_default_env().add_directive("info".parse()?))
-        .init();
+async fn main() {
+    // Required for rustls 0.23 when multiple crypto backends may be linked via deps.
+    let _ = rustls::crypto::ring::default_provider().install_default();
 
-    let endpoint = std::env::var("COSI_ENDPOINT").unwrap_or_else(|_| DEFAULT_ENDPOINT.to_string());
-    let socket_path = parse_unix_endpoint(&endpoint)?;
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
+        .with_target(true)
+        .try_init();
 
-    if let Some(parent) = socket_path.parent() {
-        tokio::fs::create_dir_all(parent).await?;
-    }
-    let _ = tokio::fs::remove_file(&socket_path).await;
-
-    let backend = BackendFactory::try_default().await?;
-    let identity = IdentityService {
-        name: DRIVER_NAME.to_string(),
+    let endpoint = std::env::var("COSI_ENDPOINT")
+        .unwrap_or_else(|_| "unix:///var/lib/cosi/cosi.sock".to_string());
+    let sock_path = match parse_unix_endpoint(&endpoint) {
+        Ok(path) => path,
+        Err(err) => {
+            error!(error = %err, "invalid COSI_ENDPOINT");
+            std::process::exit(2);
+        }
     };
-    let provisioner = ProvisionerService { backend };
 
-    let uds = UnixListener::bind(&socket_path)?;
-    let uds_stream = UnixListenerStream::new(uds);
+    if let Some(parent) = sock_path.parent() {
+        let _ = std::fs::create_dir_all(parent);
+    }
+    let _ = std::fs::remove_file(&sock_path);
+
+    let kube = match Client::try_default().await {
+        Ok(client) => client,
+        Err(err) => {
+            error!(error = %err, "failed to create Kubernetes client");
+            std::process::exit(1);
+        }
+    };
+
+    let listener = match UnixListener::bind(&sock_path) {
+        Ok(listener) => listener,
+        Err(err) => {
+            error!(error = %err, path = %sock_path.display(), "failed to bind COSI socket");
+            std::process::exit(1);
+        }
+    };
+    let incoming = UnixListenerStream::new(listener);
+    let driver = Driver::new(kube);
 
     info!(
         driver = DRIVER_NAME,
@@ -58,66 +85,14 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         "starting RustFS COSI driver"
     );
 
-    let (shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel::<()>();
-    tokio::spawn(async move {
-        wait_for_shutdown().await;
-        let _ = shutdown_tx.send(());
-    });
+    let result = tonic::transport::Server::builder()
+        .add_service(IdentityServer::new(driver.clone()))
+        .add_service(ProvisionerServer::new(driver))
+        .serve_with_incoming(incoming)
+        .await;
 
-    Server::builder()
-        .add_service(IdentityServer::new(identity))
-        .add_service(ProvisionerServer::new(provisioner))
-        .serve_with_incoming_shutdown(uds_stream, async {
-            let _ = shutdown_rx.await;
-            info!("shutdown signal received");
-        })
-        .await?;
-
-    let _ = tokio::fs::remove_file(&socket_path).await;
-    // Allow in-flight RPCs a brief window before process exit.
-    tokio::time::sleep(SHUTDOWN_GRACE).await;
-    info!("RustFS COSI driver stopped");
-    Ok(())
-}
-
-fn parse_unix_endpoint(endpoint: &str) -> Result<PathBuf, String> {
-    let endpoint = endpoint.trim();
-    if let Some(path) = endpoint.strip_prefix("unix://") {
-        if path.is_empty() {
-            return Err("COSI_ENDPOINT unix path is empty".into());
-        }
-        return Ok(PathBuf::from(path));
-    }
-    if endpoint.starts_with('/') {
-        return Ok(PathBuf::from(endpoint));
-    }
-    Err(format!(
-        "unsupported COSI_ENDPOINT `{endpoint}` (expected unix:///path/to.sock)"
-    ))
-}
-
-async fn wait_for_shutdown() {
-    let ctrl_c = async {
-        if let Err(err) = tokio::signal::ctrl_c().await {
-            warn!(error = %err, "failed to install Ctrl+C handler");
-        }
-    };
-
-    #[cfg(unix)]
-    let terminate = async {
-        match tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate()) {
-            Ok(mut stream) => {
-                stream.recv().await;
-            }
-            Err(err) => warn!(error = %err, "failed to install SIGTERM handler"),
-        }
-    };
-
-    #[cfg(not(unix))]
-    let terminate = std::future::pending::<()>();
-
-    tokio::select! {
-        () = ctrl_c => {},
-        () = terminate => {},
+    if let Err(err) = result {
+        error!(error = %err, "RustFS COSI driver stopped");
+        std::process::exit(1);
     }
 }
