--- crates/rustfs-admin/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "rustfs-admin"
-version = "0.1.0"
-edition = "2024"
-license = "Apache-2.0"
-description = "Kube-agnostic RustFS admin/S3/STS client"
-
-[dependencies]
-chrono = { version = "0.4", features = ["serde"] }
-hex = "0.4"
-hmac = "0.12"
-reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-sha2 = "0.10"
-url = "2.5"
-
-[dev-dependencies]
-axum = { version = "0.7", features = ["macros", "json"] }
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "sync", "time"] }
-
-[lints.rust]
-unused_variables = "allow"
