--- deploy/rustfs-operator/templates/cosi-deployment.yaml
@@ -1,88 +1,0 @@
-{{- if .Values.cosi.enabled }}
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: {{ include "rustfs-operator.fullname" . }}-cosi
-  namespace: {{ include "rustfs-operator.namespace" . }}
-  labels:
-    {{- include "rustfs-operator.labels" . | nindent 4 }}
-    app.kubernetes.io/component: cosi
-  {{- with .Values.commonAnnotations }}
-  annotations:
-    {{- toYaml . | nindent 4 }}
-  {{- end }}
-spec:
-  replicas: {{ .Values.cosi.replicas }}
-  selector:
-    matchLabels:
-      {{- include "rustfs-operator.selectorLabels" . | nindent 6 }}
-      app.kubernetes.io/component: cosi
-  template:
-    metadata:
-      labels:
-        {{- include "rustfs-operator.selectorLabels" . | nindent 8 }}
-        app.kubernetes.io/component: cosi
-    spec:
-      serviceAccountName: {{ include "rustfs-operator.cosiServiceAccountName" . }}
-      {{- with (.Values.cosi.imagePullSecrets | default .Values.operator.imagePullSecrets) }}
-      imagePullSecrets:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
-      {{- with .Values.cosi.podSecurityContext }}
-      securityContext:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
-      volumes:
-        - name: socket
-          emptyDir: {}
-      containers:
-        - name: rustfs-cosi-driver
-          image: {{ include "rustfs-operator.cosiImage" . }}
-          imagePullPolicy: {{ .Values.cosi.image.pullPolicy }}
-          command: ["/app/rustfs-cosi-driver"]
-          env:
-            - name: COSI_ENDPOINT
-              value: unix:///var/lib/cosi/cosi.sock
-            - name: RUST_LOG
-              value: info
-          volumeMounts:
-            - name: socket
-              mountPath: /var/lib/cosi
-          {{- with .Values.cosi.securityContext }}
-          securityContext:
-            {{- toYaml . | nindent 12 }}
-          {{- end }}
-          {{- with .Values.cosi.resources }}
-          resources:
-            {{- toYaml . | nindent 12 }}
-          {{- end }}
-        - name: objectstorage-provisioner
-          image: "{{ .Values.cosi.sidecar.image.repository }}:{{ .Values.cosi.sidecar.image.tag }}"
-          imagePullPolicy: {{ .Values.cosi.sidecar.image.pullPolicy }}
-          args:
-            - "--v=4"
-          env:
-            - name: POD_NAMESPACE
-              valueFrom:
-                fieldRef:
-                  fieldPath: metadata.namespace
-          volumeMounts:
-            - name: socket
-              mountPath: /var/lib/cosi
-          {{- with .Values.cosi.sidecar.resources }}
-          resources:
-            {{- toYaml . | nindent 12 }}
-          {{- end }}
-      {{- with .Values.cosi.nodeSelector }}
-      nodeSelector:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
-      {{- with .Values.cosi.affinity }}
-      affinity:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
-      {{- with .Values.cosi.tolerations }}
-      tolerations:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
-{{- end }}
