--- deploy/rustfs-operator/templates/cosi-rbac.yaml
@@ -1,59 +1,0 @@
-{{- if and .Values.cosi.enabled .Values.cosi.rbac.create }}
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: {{ include "rustfs-operator.fullname" . }}-cosi
-  labels:
-    {{- include "rustfs-operator.labels" . | nindent 4 }}
-    app.kubernetes.io/component: cosi
-  {{- with .Values.commonAnnotations }}
-  annotations:
-    {{- toYaml . | nindent 4 }}
-  {{- end }}
-rules:
-  # Driver reads Tenant admin Secrets and optional TLS CA ConfigMaps referenced by BucketClass parameters.
-  - apiGroups: [""]
-    resources: ["secrets", "configmaps"]
-    verbs: ["get", "list", "watch"]
-  # Sidecar reconciles COSI API objects (v1alpha1).
-  - apiGroups: ["objectstorage.k8s.io"]
-    resources:
-      - buckets
-      - bucketaccesses
-      - bucketclaims
-      - bucketaccessclasses
-      - bucketclasses
-    verbs: ["get", "list", "watch", "update", "patch", "create", "delete"]
-  - apiGroups: ["objectstorage.k8s.io"]
-    resources:
-      - buckets/status
-      - bucketaccesses/status
-      - bucketclaims/status
-    verbs: ["get", "update", "patch"]
-  - apiGroups: ["coordination.k8s.io"]
-    resources: ["leases"]
-    verbs: ["get", "list", "watch", "create", "update", "patch"]
-  - apiGroups: [""]
-    resources: ["events"]
-    verbs: ["create", "patch", "update"]
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: {{ include "rustfs-operator.fullname" . }}-cosi
-  labels:
-    {{- include "rustfs-operator.labels" . | nindent 4 }}
-    app.kubernetes.io/component: cosi
-  {{- with .Values.commonAnnotations }}
-  annotations:
-    {{- toYaml . | nindent 4 }}
-  {{- end }}
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: {{ include "rustfs-operator.fullname" . }}-cosi
-subjects:
-  - kind: ServiceAccount
-    name: {{ include "rustfs-operator.cosiServiceAccountName" . }}
-    namespace: {{ include "rustfs-operator.namespace" . }}
-{{- end }}
