--- deploy/rustfs-operator/templates/cosi-serviceaccount.yaml
@@ -1,21 +1,0 @@
-{{- if .Values.cosi.enabled }}
-{{- if .Values.cosi.serviceAccount.create -}}
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: {{ include "rustfs-operator.cosiServiceAccountName" . }}
-  namespace: {{ include "rustfs-operator.namespace" . }}
-  labels:
-    {{- include "rustfs-operator.labels" . | nindent 4 }}
-    app.kubernetes.io/component: cosi
-  {{- if or .Values.cosi.serviceAccount.annotations .Values.commonAnnotations }}
-  annotations:
-    {{- with .Values.commonAnnotations }}
-    {{- toYaml . | nindent 4 }}
-    {{- end }}
-    {{- with .Values.cosi.serviceAccount.annotations }}
-    {{- toYaml . | nindent 4 }}
-    {{- end }}
-  {{- end }}
-{{- end }}
-{{- end }}
