+++ deploy/rustfs-operator/values.yaml
@@ -307,67 +307,3 @@ console:
       # - secretName: console-tls
       #   hosts:
       #     - console.example.com
-
-# COSI (Container Object Storage Interface) driver — experimental, v1alpha1.
-# Requires the COSI controller/CRDs installed separately:
-#   kubectl apply -k 'github.com/kubernetes-sigs/container-object-storage-interface//?ref=release-0.2'
-cosi:
-  enabled: false
-
-  driverName: rustfs.objectstorage.k8s.io
-
-  replicas: 1
-
-  # Driver binary image (defaults to the operator image which embeds rustfs-cosi-driver).
-  image:
-    repository: ""  # defaults to operator.image.repository
-    tag: ""         # defaults to operator.image.tag
-    pullPolicy: IfNotPresent
-
-  imagePullSecrets: []
-
-  # Official COSI provisioner sidecar (UNIX socket to the driver).
-  sidecar:
-    image:
-      repository: gcr.io/k8s-staging-sig-storage/objectstorage-sidecar
-      tag: v20230130-v0.1.0-24-gc0cf995
-      pullPolicy: IfNotPresent
-    resources:
-      requests:
-        cpu: 10m
-        memory: 32Mi
-      limits:
-        memory: 256Mi
-
-  resources:
-    requests:
-      cpu: 50m
-      memory: 64Mi
-    limits:
-      cpu: 500m
-      memory: 256Mi
-
-  podSecurityContext:
-    fsGroup: 65534
-
-  securityContext:
-    allowPrivilegeEscalation: false
-    runAsNonRoot: true
-    runAsUser: 65534
-    capabilities:
-      drop:
-        - ALL
-    seccompProfile:
-      type: RuntimeDefault
-
-  nodeSelector: {}
-  tolerations: []
-  affinity: {}
-
-  serviceAccount:
-    create: true
-    annotations: {}
-    name: ""
-
-  rbac:
-    create: true
