--- examples/cosi/bucketaccessclass.yaml
@@ -1,13 +1,0 @@
-apiVersion: objectstorage.k8s.io/v1alpha1
-kind: BucketAccessClass
-metadata:
-  name: rustfs-bucket-access-class
-driverName: rustfs.objectstorage.k8s.io
-authenticationType: KEY
-parameters:
-  objectStoreUserSecretName: provisioning-admin-creds
-  objectStoreUserSecretNamespace: default
-  endpoint: http://provisioning-demo-io.default.svc:9000
-  region: us-east-1
-  # readonly | readwrite (default)
-  policy: readwrite
