+++ src/reconcile/pool_lifecycle.rs
@@ -25,7 +25,7 @@ use super::{Error, context};
 use crate::context::Context;
 use crate::sts::rustfs_client::{
     RustfsAdminClient, RustfsClientError, RustfsPoolDecommissionInfo, RustfsPoolListItem,
-    RustfsPoolStatus, client_from_tenant, client_from_tls_tenant_for_sts, load_tenant_credentials,
+    RustfsPoolStatus,
 };
 use crate::types::v1alpha1::pool::Pool;
 use crate::types::v1alpha1::pool_lifecycle::{DecommissionAction, DecommissionRequest};
@@ -453,11 +453,17 @@ async fn rustfs_admin_client(
     ctx: &Context,
     tenant: &Tenant,
 ) -> Result<RustfsAdminClient, RustfsClientError> {
-    let credentials = load_tenant_credentials(&ctx.client, tenant).await?;
+    let credentials = RustfsAdminClient::load_tenant_credentials(&ctx.client, tenant).await?;
     if tenant.spec.tls.as_ref().is_some_and(|tls| tls.is_enabled()) {
-        client_from_tls_tenant_for_sts(&ctx.client, tenant, credentials, ctx.cluster_domain()).await
+        RustfsAdminClient::from_tls_tenant_for_sts(
+            &ctx.client,
+            tenant,
+            credentials,
+            ctx.cluster_domain(),
+        )
+        .await
     } else {
-        client_from_tenant(tenant, credentials)
+        RustfsAdminClient::from_tenant(tenant, credentials)
     }
 }
