+++ src/sts/admin_ops.rs
@@ -18,18 +18,44 @@
 
 use std::collections::BTreeMap;
 
+use super::helpers::{
+    body_mentions_not_found, build_canonical_query, extract_canned_policy_document,
+};
+use super::{
+    ADD_CANNED_POLICY_PATH, ADD_USER_PATH, ADMIN_SIGNING_SERVICE, INFO_CANNED_POLICY_PATH,
+    JSON_CONTENT_TYPE, LIST_CANNED_POLICIES_PATH, REMOVE_USER_PATH, RustfsAdminClient,
+    RustfsClientError, RustfsServerInfo, RustfsServerInfoResponse, RustfsUserInfo,
+    SERVER_INFO_PATH, SET_POLICY_PATH, USER_INFO_PATH,
+};
 use reqwest::StatusCode;
 use serde_json::Value;
 
-use crate::client::{
-    ADD_CANNED_POLICY_PATH, ADD_USER_PATH, ADMIN_SIGNING_SERVICE, INFO_CANNED_POLICY_PATH,
-    JSON_CONTENT_TYPE, LIST_CANNED_POLICIES_PATH, REMOVE_CANNED_POLICY_PATH, REMOVE_USER_PATH,
-    RustfsAdminClient, RustfsClientError, RustfsServerInfo, RustfsServerInfoResponse,
-    SERVER_INFO_PATH, SET_POLICY_PATH, USER_INFO_PATH,
-};
-use crate::helpers::{
-    body_mentions_not_found, build_canonical_query, extract_canned_policy_document,
-};
+fn parse_user_info_policy_names(body: &Value) -> Vec<String> {
+    let Some(field) = body
+        .get("policyName")
+        .or_else(|| body.get("policy_name"))
+        .or_else(|| body.get("PolicyName"))
+    else {
+        return Vec::new();
+    };
+
+    match field {
+        Value::String(raw) => raw
+            .split(',')
+            .map(str::trim)
+            .filter(|name| !name.is_empty())
+            .map(ToOwned::to_owned)
+            .collect(),
+        Value::Array(items) => items
+            .iter()
+            .filter_map(|item| item.as_str())
+            .map(str::trim)
+            .filter(|name| !name.is_empty())
+            .map(ToOwned::to_owned)
+            .collect(),
+        _ => Vec::new(),
+    }
+}
 
 impl RustfsAdminClient {
     // Admin duties: user and policy management APIs.
@@ -122,45 +148,6 @@ impl RustfsAdminClient {
         Ok(())
     }
 
-    /// Remove a RustFS canned policy. Succeeds if the policy is already gone.
-    pub async fn remove_canned_policy(&self, policy_name: &str) -> Result<(), RustfsClientError> {
-        if policy_name.trim().is_empty() {
-            return Err(RustfsClientError::InvalidPolicyName);
-        }
-
-        let query = build_canonical_query(&[("name", policy_name)]);
-        let path = REMOVE_CANNED_POLICY_PATH;
-        let url = format!("{}{}?{query}", self.base_url.trim_end_matches('/'), path);
-
-        let signed = self.sign_request("DELETE", path, &query, "", None, ADMIN_SIGNING_SERVICE)?;
-        let host = self.host()?;
-
-        let response = self
-            .http_client
-            .delete(url)
-            .header("x-amz-date", &signed.amz_date)
-            .header("x-amz-content-sha256", &signed.payload_hash)
-            .header("authorization", &signed.authorization)
-            .header("host", host)
-            .send()
-            .await
-            .map_err(|_| RustfsClientError::RequestFailed)?;
-
-        if response.status().is_success() {
-            return Ok(());
-        }
-
-        let status = response.status();
-        let (body, truncated) = RustfsClientError::limited_response_body(response).await;
-        if status == StatusCode::NOT_FOUND || body_mentions_not_found(&body) {
-            return Ok(());
-        }
-
-        Err(RustfsClientError::unexpected_status_with_limited_body(
-            status, &body, truncated,
-        ))
-    }
-
     pub async fn list_canned_policies(
         &self,
     ) -> Result<BTreeMap<String, String>, RustfsClientError> {
@@ -196,7 +183,11 @@ impl RustfsAdminClient {
             .map_err(|_| RustfsClientError::ParseResponseFailed)
     }
 
-    pub async fn user_exists(&self, access_key: &str) -> Result<bool, RustfsClientError> {
+    /// Fetch IAM user info. Returns `Ok(None)` when the user does not exist.
+    pub async fn get_user_info(
+        &self,
+        access_key: &str,
+    ) -> Result<Option<RustfsUserInfo>, RustfsClientError> {
         if access_key.trim().is_empty() {
             return Err(RustfsClientError::InvalidCredentialValue { key: "accesskey" });
         }
@@ -219,18 +210,30 @@ impl RustfsAdminClient {
             .map_err(|_| RustfsClientError::RequestFailed)?;
 
         if response.status().is_success() {
-            return Ok(true);
+            let body = response
+                .text()
+                .await
+                .map_err(|_| RustfsClientError::RequestFailed)?;
+            let parsed: Value =
+                serde_json::from_str(&body).map_err(|_| RustfsClientError::ParseResponseFailed)?;
+            return Ok(Some(RustfsUserInfo {
+                policy_names: parse_user_info_policy_names(&parsed),
+            }));
         }
 
         let status = response.status();
         let (body, truncated) = RustfsClientError::limited_response_body(response).await;
         if status == StatusCode::NOT_FOUND || body_mentions_not_found(&body) {
-            return Ok(false);
+            return Ok(None);
         }
 
         Err(RustfsClientError::unexpected_status_with_limited_body(
             status, &body, truncated,
         ))
+    }
+
+    pub async fn user_exists(&self, access_key: &str) -> Result<bool, RustfsClientError> {
+        Ok(self.get_user_info(access_key).await?.is_some())
     }
 
     pub async fn add_user(
@@ -257,7 +260,31 @@ impl RustfsAdminClient {
             .map(|_| ())
     }
 
-    /// Remove a RustFS user. Succeeds if the user is already gone.
+    pub async fn set_user_policy(
+        &self,
+        access_key: &str,
+        policies: &[String],
+    ) -> Result<(), RustfsClientError> {
+        if access_key.trim().is_empty() {
+            return Err(RustfsClientError::InvalidCredentialValue { key: "accesskey" });
+        }
+        if policies.is_empty() || policies.iter().any(|policy| policy.trim().is_empty()) {
+            return Err(RustfsClientError::InvalidPolicyName);
+        }
+
+        let policy_names = policies.join(",");
+        let query = build_canonical_query(&[
+            ("isGroup", "false"),
+            ("policyName", policy_names.as_str()),
+            ("userOrGroup", access_key),
+        ]);
+
+        self.send_admin_request("PUT", SET_POLICY_PATH, &query, "", None)
+            .await
+            .map(|_| ())
+    }
+
+    /// Remove a RustFS user. Missing users are treated as success (idempotent).
     pub async fn remove_user(&self, access_key: &str) -> Result<(), RustfsClientError> {
         if access_key.trim().is_empty() {
             return Err(RustfsClientError::InvalidCredentialValue { key: "accesskey" });
@@ -266,7 +293,6 @@ impl RustfsAdminClient {
         let query = build_canonical_query(&[("accessKey", access_key)]);
         let path = REMOVE_USER_PATH;
         let url = format!("{}{}?{query}", self.base_url.trim_end_matches('/'), path);
-
         let signed = self.sign_request("DELETE", path, &query, "", None, ADMIN_SIGNING_SERVICE)?;
         let host = self.host()?;
 
@@ -295,28 +321,33 @@ impl RustfsAdminClient {
             status, &body, truncated,
         ))
     }
+}
 
-    pub async fn set_user_policy(
-        &self,
-        access_key: &str,
-        policies: &[String],
-    ) -> Result<(), RustfsClientError> {
-        if access_key.trim().is_empty() {
-            return Err(RustfsClientError::InvalidCredentialValue { key: "accesskey" });
-        }
-        if policies.is_empty() || policies.iter().any(|policy| policy.trim().is_empty()) {
-            return Err(RustfsClientError::InvalidPolicyName);
-        }
+#[cfg(test)]
+mod parse_tests {
+    use super::parse_user_info_policy_names;
+    use serde_json::json;
 
-        let policy_names = policies.join(",");
-        let query = build_canonical_query(&[
-            ("isGroup", "false"),
-            ("policyName", policy_names.as_str()),
-            ("userOrGroup", access_key),
-        ]);
+    #[test]
+    fn parses_comma_separated_policy_name() {
+        let body = json!({"policyName":"cosi-mlflow,cosi-grant-ba-1"});
+        assert_eq!(
+            parse_user_info_policy_names(&body),
+            vec!["cosi-mlflow".to_string(), "cosi-grant-ba-1".to_string()]
+        );
+    }
 
-        self.send_admin_request("PUT", SET_POLICY_PATH, &query, "", None)
-            .await
-            .map(|_| ())
+    #[test]
+    fn parses_policy_name_array_and_snake_case() {
+        let body = json!({"policy_name":["a","b"]});
+        assert_eq!(
+            parse_user_info_policy_names(&body),
+            vec!["a".to_string(), "b".to_string()]
+        );
+    }
+
+    #[test]
+    fn missing_policy_field_yields_empty() {
+        assert!(parse_user_info_policy_names(&json!({"status":"enabled"})).is_empty());
     }
 }
